+++ src/Controller/AuthBaseController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace CaliforniaMountainSnake\SimpleLaravelAuthSystem\Controller;
+
+use Illuminate\Foundation\Http\Middleware\ConvertEmptyStringsToNull;
+use Illuminate\Foundation\Http\Middleware\TrimStrings;
+
+class AuthBaseController extends ClearLaravelController
+{
+    public function __construct()
+    {
+        $this->middleware(TrimStrings::class);
+        $this->middleware(ConvertEmptyStringsToNull::class);
+    }
+}